@@ -11,15 +11,13 @@ const mix = require('laravel-mix')
  |
  */
 
+mix.sass('resources/sass/app.scss', 'public/css').sourceMaps(false, 'source-map')
+mix.js('resources/js/app.js', 'public/js').vue()
 
-mix.sass('resources/sass/app.scss', 'public/css').sourceMaps(false, 'source-map');
-mix.js('resources/js/app.js', 'public/js');
-
-mix.sass('resources/sass/admin/admin.scss', 'public/css').sourceMaps(false, 'source-map');
-mix.js('resources/js/admin.js', 'public/js');
-
+mix.sass('resources/sass/admin/admin.scss', 'public/css').sourceMaps(false, 'source-map')
+// mix.js('resources/js/admin.js', 'public/js').vue();
 
 // cache busting in production
-if (mix.inProduction()) {
-  mix.version();
-}
+// if (mix.inProduction()) {
+//  mix.version();
+// }
